Optimize name transformations
These can be run at compile time. Previously they were run at runtime on
every invocation of the codec. This was particularly expensive for some
of the builtins like the snake_case codec as it was allocating and
compiling a Pattern instance per member field per invocation of the
codec